Conflicts:
	Readme.md

+++ info/views.py
@@ -7,8 +7,9 @@ def info(request):
     
     if request.method == "POST":
         try:
-            lat = int(request.POST["lat"])
-            lon = int(request.POST["lon"])
+            # you might want to use Number to prevent having a million zeros
+            lat = float(request.POST["lat"])
+            lon = float(request.POST["lon"])
             return HttpResponse(content = "some information", content_type = "text/json")
         except Exception, e:
             return HttpResponse(content = "Something went wrong", content_type = "text/json")
